pyproject.toml
[tool.poetry]
name = "yapping_pandas"
version = "0.1.0"
description = ""
authors = ["Julian Streit <julian.streit@tum.de>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
flask = "^2.2.2"
sqlalchemy = "^2.0.0"
geoalchemy2 = "^0.12.5"
pydantic = "^2.0.0"
pure-protobuf = "^2.1.1"
shapely = "^2.0.0"
flask-pydantic = "^0.11.0"
tornado = "^6.2"
fastapi = "^0.115.0"
singleton-decorator = "^1.0.0"
uvicorn = "^0.19.0"
geojson-pydantic = "^1.0.0"
psycopg2-binary = "^2.9.5"
pymongo = "^4.3.3"
requests = "^2.28.1"
sqlalchemy-utils = "^0.41.2"
geopandas = "^1.0.0"
rasterio = "1.3.4"
scipy = "^1.10.1"
tqdm = "^4.65.0"
mercantile = "^1.2.1"
pillow = "^9.5.0"
alembic = "^1.13.2"
alembic-utils = "^0.8.4"
alembic-postgresql-enum = "^1.3.0"
pydantic-settings = "^2.5.2"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
mypy = "^0.991"
pytest = "^7.2.0"
pytest-mock = "^3.10.0"
pytest-cov = "^4.0.0"
httpx = "^0.23.1"
flake8 = "^6.0.0"
coverage = "^7.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
